refactor: split Field types into separate file
