+++ packages/core/components/AutoField/fields/DefaultField/index.tsx
@@ -1,7 +1,8 @@
 import getClassNameFactory from "../../../../lib/get-class-name-factory";
 import styles from "../../styles.module.css";
 import { Hash, Type } from "lucide-react";
-import { FieldLabelInternal, type InputProps } from "../..";
+import { FieldLabelInternal } from "../..";
+import type { FieldProps } from "../../../../types/Fields";
 
 const getClassName = getClassNameFactory("Input", styles);
 
@@ -13,7 +14,7 @@ export const DefaultField = ({
   name,
   label,
   id,
-}: InputProps) => {
+}: FieldProps) => {
   return (
     <FieldLabelInternal
       label={label || name}
