+++ packages/core/components/AutoField/fields/RadioField/index.tsx
@@ -1,7 +1,8 @@
 import getClassNameFactory from "../../../../lib/get-class-name-factory";
 import styles from "../../styles.module.css";
 import { CheckCircle } from "lucide-react";
-import { FieldLabelInternal, type InputProps } from "../..";
+import { FieldLabelInternal } from "../..";
+import type { FieldProps } from "../../../../types/Fields";
 
 const getClassName = getClassNameFactory("Input", styles);
 
@@ -13,7 +14,7 @@ export const RadioField = ({
   name,
   id,
   label,
-}: InputProps) => {
+}: FieldProps) => {
   if (field.type !== "radio" || !field.options) {
     return null;
   }
