+++ packages/core/index.ts
@@ -3,6 +3,7 @@ import "./styles.css";
 export type { PuckAction } from "./reducer/actions";
 
 export * from "./types/Config";
+export * from "./types/Fields";
 export * from "./components/Button";
 export { Drawer } from "./components/Drawer";
