+++ packages/core/types/Overrides.ts
@@ -1,6 +1,5 @@
 import { ReactElement, ReactNode } from "react";
-import { InputProps } from "../components/AutoField";
-import { Field } from "./Config";
+import { Field, FieldProps } from "./Fields";
 import { ItemSelector } from "../lib/get-item";
 
 // Plugins can use `usePuck` instead of relying on props
@@ -52,7 +51,7 @@ export type Overrides = OverridesGeneric<{
 export type FieldRenderFunctions = Omit<
   {
     [Type in Field["type"]]: React.FunctionComponent<
-      InputProps<Extract<Field, { type: Type }>> & {
+      FieldProps<Extract<Field, { type: Type }>> & {
         children: ReactNode;
       }
     >;
@@ -60,7 +59,7 @@ export type FieldRenderFunctions = Omit<
   "custom"
 > & {
   [key: string]: React.FunctionComponent<
-    InputProps<any> & {
+    FieldProps<any> & {
       children: ReactNode;
     }
   >;
